+++ examples/09-flush-to-persistent-GPSPM/client.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /* Copyright 2020-2022, Intel Corporation */
-/* Copyright 2021, Fujitsu */
+/* Copyright 2021-2022, Fujitsu */
 
 /*
  * client.c -- a client of the flush-to-persistent-GPSPM example
@@ -70,7 +70,7 @@ main(int argc, char *argv[])
 	size_t dst_size = 0;
 	size_t dst_offset = 0;
 	struct rpma_mr_local *src_mr = NULL;
-	struct rpma_completion cmpl;
+	struct ibv_wc wc;
 
 	/* messaging resources */
 	void *msg_ptr = NULL;
@@ -269,24 +269,24 @@ main(int argc, char *argv[])
 		goto err_mr_remote_delete;
 	if ((ret = rpma_cq_wait(cq)))
 		goto err_mr_remote_delete;
-	if ((ret = rpma_cq_get_completion(cq, &cmpl)))
+	if ((ret = rpma_cq_get_wc(cq, 1, &wc, NULL)))
 		goto err_mr_remote_delete;
 
 	/* validate the send completion */
-	if (cmpl.op_status != IBV_WC_SUCCESS) {
+	if (wc.status != IBV_WC_SUCCESS) {
 		ret = -1;
 		(void) fprintf(stderr, "rpma_send() failed: %s\n",
-				ibv_wc_status_str(cmpl.op_status));
+				ibv_wc_status_str(wc.status));
 		goto err_mr_remote_delete;
 	}
 
-	if (cmpl.op != RPMA_OP_SEND) {
+	if (wc.opcode != IBV_WC_SEND) {
 		ret = -1;
 		(void) fprintf(stderr,
-				"unexpected cmpl.op value "
+				"unexpected wc.opcode value "
 				"(0x%" PRIXPTR " != 0x%" PRIXPTR ")\n",
-				(uintptr_t)cmpl.op,
-				(uintptr_t)RPMA_OP_SEND);
+				(uintptr_t)wc.opcode,
+				(uintptr_t)IBV_WC_SEND);
 		goto err_mr_remote_delete;
 	}
 
@@ -296,29 +296,29 @@ main(int argc, char *argv[])
 		goto err_mr_remote_delete;
 	if ((ret = rpma_cq_wait(rcq)))
 		goto err_mr_remote_delete;
-	if ((ret = rpma_cq_get_completion(rcq, &cmpl)))
+	if ((ret = rpma_cq_get_wc(rcq, 1, &wc, NULL)))
 		goto err_mr_remote_delete;
 
 	/* validate the receive completion */
-	if (cmpl.op_status != IBV_WC_SUCCESS) {
+	if (wc.status != IBV_WC_SUCCESS) {
 		ret = -1;
 		(void) fprintf(stderr, "rpma_recv() failed: %s\n",
-				ibv_wc_status_str(cmpl.op_status));
+				ibv_wc_status_str(wc.status));
 		goto err_mr_remote_delete;
 	}
 
-	if (cmpl.op != RPMA_OP_RECV) {
+	if (wc.opcode != IBV_WC_RECV) {
 		ret = -1;
 		(void) fprintf(stderr,
-				"unexpected cmpl.op value "
+				"unexpected wc.opcode value "
 				"(0x%" PRIXPTR " != 0x%" PRIXPTR ")\n",
-				(uintptr_t)cmpl.op,
-				(uintptr_t)RPMA_OP_RECV);
+				(uintptr_t)wc.opcode,
+				(uintptr_t)IBV_WC_RECV);
 		goto err_mr_remote_delete;
 	}
 
 	/* unpack a response from the received buffer */
-	flush_resp = gpspm_flush_response__unpack(NULL, cmpl.byte_len,
+	flush_resp = gpspm_flush_response__unpack(NULL, wc.byte_len,
 			recv_ptr);
 	if (flush_resp == NULL) {
 		fprintf(stderr, "Cannot unpack the flush response buffer\n");